// backend/src/config/env.d.ts
export declare const env: {
    PORT: string;
    NODE_ENV: "development" | "production" | "test";
    FIREBASE_PROJECT_ID?: string | undefined;
    FIREBASE_CLIENT_EMAIL?: string | undefined;
    FIREBASE_PRIVATE_KEY?: string | undefined;
    GEMINI_API_KEY?: string | undefined;
};
//# sourceMappingURL=env.d.ts.map